requirements.txt
aiogram>=3.7,<4.0
python-dotenv>=1.0
sqlalchemy>=2.0
alembic>=1.13
aiosqlite>=0.19.0
aiomysql>=0.2.0
asyncpg>=0.29.0
pydantic>=2.0
pillow>=10.0
aiohttp>=3.9
redis>=5.0
python-telegram-bot>=20.0
